requirements.txt
# Luna Gaming Tool Dependencies
rich>=13.0.0
aiohttp>=3.8.0
python-socketio>=5.7.0
psutil>=5.9.0